requirements.txt
Pillow==5.4.1
Flask==1.0.2
gunicorn==19.9.0